Bug 1284079   Need option to disable interop support in config file.
control.c 		- Added enableInterOp boolean control option with default true value
providerRegister.c 	- Changed type of exFlags from int to unsigned long
sfcBroker.c		- Added code to deal with enableInterOp control option
sfcb.cfg.pre.in		- Added "enableInterOp:  true" to config file
main/sfcbd.1.pre.in	- Documented enableInterOp config option in man page

+++ providerRegister.c
@@ -32,7 +32,7 @@
 #include "providerRegister.h"
 
 static ProviderInfo forceNotFound={"",FORCE_PROVIDER_NOTFOUND};
-extern int exFlags;
+extern unsigned long exFlags;
 extern char * configfile;
 extern int setupControl(char *fn);
 extern int getControlChars(char *id, char **val);
